Restructured Month Heirarchy
Moved files from the src/Data/Month/En/Us/ folder to the
src/Data/Region/En/Us/Month folder. This should help keep all
information for a given region together as there are new ‘helper’
classes coming down the pike.
