+++ src/Data/Region/En/Us/Month/November.php
@@ -1,5 +1,5 @@
 <?php
-namespace Skybluesofa\OnThisDay\Data\Month\En\Us;
+namespace Skybluesofa\OnThisDay\Data\Region\En\Us\Month;
 
 use Skybluesofa\OnThisDay\Data\Contract\Month;
